refactor(macos): extract ImageURLs from AppModel into AppModel+ImageURLs.swift
